+++ ChangeLog.md
@@ -17,6 +17,7 @@ Incoming in 0.9.2
 * Other
     * Renamed `CompileSafeRegexException` to `MalformedPatternException`, so it better describes its nature
     * Renamed `is()->delimitered()` to `is()->delimited()`
+    * Returning `Match` from `replace()->callback()` (instead of `Match.text()` as `string`)
 
 * Maintenance
     * PhpUnit throws different exceptions because of [PHP `__toString()` exception policy change](https://wiki.php.net/rfc/tostring_exceptions).